ARROW-7650: [C++][Dataset] enable dataset tests on Windows
Closes #6260 from bkietz/ARROW-7650-dataset-tests-on-windows and squashes the following commits:

ffc353d54 <Antoine Pitrou> Lint & fix style issues
db4c1f16d <Benjamin Kietzman> ARROW-7650:  enable dataset tests on Windows

Lead-authored-by: Benjamin Kietzman <bengilgit@gmail.com>

+++ cpp/src/arrow/dataset/scanner_test.cc
@@ -16,6 +16,7 @@
 // under the License.
 
 #include "arrow/dataset/scanner.h"
+
 #include <memory>
 
 #include "arrow/compute/context.h"
@@ -35,12 +36,13 @@ class TestScanner : public DatasetFixtureMixin {
   static constexpr int64_t kBatchSize = 1024;
 
   Scanner MakeScanner(std::shared_ptr<RecordBatch> batch) {
-    std::vector<std::shared_ptr<RecordBatch>> batches{kNumberBatches, batch};
+    std::vector<std::shared_ptr<RecordBatch>> batches{static_cast<size_t>(kNumberBatches),
+                                                      batch};
 
-    FragmentVector fragments{kNumberFragments,
+    FragmentVector fragments{static_cast<size_t>(kNumberFragments),
                              std::make_shared<InMemoryFragment>(batches, options_)};
 
-    SourceVector sources{kNumberSources,
+    SourceVector sources{static_cast<size_t>(kNumberSources),
                          std::make_shared<InMemorySource>(batch->schema(), fragments)};
 
     return Scanner{sources, options_, ctx_};
